Separate multiple description lines from Easybank with blanks
These lines are separated in the online banking with BRs, but these
are stripped by the text() routine by default. Need to select
them manually with xpath as no other functions seem to be
available.

Issue was that some of my transactions could not be parsed correctly
because the IBAN or other details were not separated from other
transaction details and the MoneyMoney parser did not detect those
correctly.